Decouple model tests from pyspark dependency
